@@ -1,55 +1,55 @@
 {
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "6.1.1",
-    "@fortawesome/free-regular-svg-icons": "6.1.1",
-    "@fortawesome/free-solid-svg-icons": "6.1.1",
-    "@fortawesome/react-fontawesome": "0.1.18",
-    "@reduxjs/toolkit": "1.8.2",
-    "autoprefixer": "10.4.7",
+    "@fortawesome/fontawesome-svg-core": "6.4.0",
+    "@fortawesome/free-regular-svg-icons": "6.4.0",
+    "@fortawesome/free-solid-svg-icons": "6.4.0",
+    "@fortawesome/react-fontawesome": "0.2.0",
+    "@reduxjs/toolkit": "1.9.3",
+    "autoprefixer": "10.4.14",
     "history": "5.3.0",
-    "react": "18.1.0",
-    "react-bootstrap": "2.4.0",
-    "react-dom": "18.1.0",
-    "react-dropzone": "14.2.1",
-    "react-i18next": "11.16.9",
-    "react-redux": "8.0.2",
-    "react-router": "6.3.0",
-    "react-router-dom": "6.3.0",
-    "react-select": "5.3.2",
-    "semantic-ui-react": "2.1.3"
+    "react": "18.2.0",
+    "react-bootstrap": "2.7.2",
+    "react-dom": "18.2.0",
+    "react-dropzone": "14.2.3",
+    "react-i18next": "12.2.0",
+    "react-redux": "8.0.5",
+    "react-router": "6.10.0",
+    "react-router-dom": "6.10.0",
+    "react-select": "5.7.2",
+    "semantic-ui-react": "2.1.4"
   },
   "devDependencies": {
-    "@babel/preset-env": "7.18.2",
-    "@babel/preset-react": "7.17.12",
-    "@babel/preset-typescript": "7.17.12",
-    "@types/react": "18.0.10",
-    "@types/react-dom": "18.0.5",
-    "@types/react-redux": "7.1.24",
-    "@typescript-eslint/eslint-plugin": "5.27.0",
-    "@typescript-eslint/parser": "5.27.0",
-    "babel-loader": "8.2.5",
-    "browserslist": "4.20.3",
-    "css-loader": "6.7.1",
-    "eslint": "8.16.0",
+    "@babel/preset-env": "7.21.4",
+    "@babel/preset-react": "7.18.6",
+    "@babel/preset-typescript": "7.21.4",
+    "@types/react": "18.0.32",
+    "@types/react-dom": "18.0.11",
+    "@types/react-redux": "7.1.25",
+    "@typescript-eslint/eslint-plugin": "5.57.0",
+    "@typescript-eslint/parser": "5.57.0",
+    "babel-loader": "9.1.2",
+    "browserslist": "4.21.5",
+    "css-loader": "6.7.3",
+    "eslint": "8.37.0",
     "eslint-config-airbnb": "19.0.4",
-    "eslint-config-prettier": "8.5.0",
-    "eslint-import-resolver-typescript": "2.7.1",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-prettier": "4.0.0",
-    "eslint-plugin-react": "7.30.0",
-    "eslint-plugin-react-hooks": "4.5.0",
-    "mini-css-extract-plugin": "2.6.0",
-    "postcss": "8.4.14",
-    "postcss-loader": "7.0.0",
-    "sass": "1.52.1",
-    "sass-loader": "13.0.0",
-    "style-loader": "3.3.1",
-    "stylelint": "14.8.5",
-    "stylelint-config-standard-scss": "4.0.0",
-    "webpack": "5.72.1",
-    "webpack-cli": "4.9.2",
-    "webpack-dev-server": "4.9.1"
+    "eslint-config-prettier": "8.8.0",
+    "eslint-import-resolver-typescript": "3.5.4",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-react": "7.32.2",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "mini-css-extract-plugin": "2.7.5",
+    "postcss": "8.4.21",
+    "postcss-loader": "7.1.0",
+    "sass": "1.60.0",
+    "sass-loader": "13.2.2",
+    "style-loader": "3.3.2",
+    "stylelint": "15.4.0",
+    "stylelint-config-standard-scss": "7.0.1",
+    "webpack": "5.77.0",
+    "webpack-cli": "5.0.1",
+    "webpack-dev-server": "4.13.2"
   },
   "browserslist": [
     "defaults",
