@@ -1,5 +1,5 @@
 import * as React from "react"
-import * as ReactDOM from "react-dom"
+import * as ReactDOMClient from "react-dom/client"
 import { BrowserRouter as Router, Link, Route, Routes } from "react-router-dom"
 import { library } from "@fortawesome/fontawesome-svg-core"
 import { fas } from "@fortawesome/free-solid-svg-icons"
@@ -158,7 +158,8 @@ const store = configureStore({ reducer: rootReducer })
 
 library.add(fas, far)
 
-ReactDOM.render(
+const root = ReactDOMClient.createRoot(document.getElementById("root"))
+root.render(
   <Provider store={store}>
     <Router>
       <Base>
@@ -175,5 +176,4 @@ ReactDOM.render(
       </Base>
     </Router>
   </Provider>,
-  document.getElementById("root"),
 )
